requirements.txt.txt
python-telegram-bot==20.7
Flask==2.3.3
python-dotenv==1.0.0
aiohttp==3.9.0
requests==2.31.0